+++ android/build.gradle
@@ -2,7 +2,7 @@ group 'live.ditto.ditto_flutter_tools'
 version '1.0.0'
 
 buildscript {
-    ext.kotlin_version = '1.7.10'
+    ext.kotlin_version = '1.9.20'
     repositories {
         google()
         mavenCentral()
