+++ example/android/app/build.gradle
@@ -24,7 +24,7 @@ if (flutterVersionName == null) {
 }
 
 android {
-    namespace = "com.example.example"
+    namespace = "com.ditto.example"
     compileSdk = 35 
     ndkVersion = "27.0.12077973"
