+++ example/android/app/src/main/kotlin/com/ditto/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example
+package com.ditto.example
 
 import io.flutter.embedding.android.FlutterActivity
