+++ lib/src/permissions_health/wifi_permissions_stub.dart
@@ -1,30 +1,6 @@
-Never get _unsupportedPlatformStub => throw "Unsupported platform for WiFi permissions checking";
+import 'wifi_permissions_types.dart';
 
-enum WifiPermissionStatus {
-  enabled,
-  notConfigured,
-  notSupported,
-}
-
-class WifiPermissionResult {
-  final WifiPermissionStatus status;
-  final String message;
-
-  const WifiPermissionResult({
-    required this.status,
-    required this.message,
-  });
-}
-
-class WifiAwarePermissionResult {
-  final WifiPermissionStatus status;
-  final String message;
-
-  const WifiAwarePermissionResult({
-    required this.status,
-    required this.message,
-  });
-}
+Never get _unsupportedPlatformStub => throw "Unsupported Platform";
 
 /// Check if WiFi permissions are properly configured for Ditto
 Future<WifiPermissionResult> checkWifiPermissions() => _unsupportedPlatformStub;
