+++ lib/src/permissions_health/wifi_permissions_types.dart
@@ -1,0 +1,26 @@
+/// Shared types for WiFi permissions health checking
+enum WifiPermissionStatus {
+  enabled,
+  notConfigured,
+  notSupported,
+}
+
+class WifiPermissionResult {
+  final WifiPermissionStatus status;
+  final String message;
+
+  const WifiPermissionResult({
+    required this.status,
+    required this.message,
+  });
+}
+
+class WifiAwarePermissionResult {
+  final WifiPermissionStatus status;
+  final String message;
+
+  const WifiAwarePermissionResult({
+    required this.status,
+    required this.message,
+  });
+}
