+++ lib/src/permissions_health/wifi_permissions_web.dart
@@ -1,28 +1,4 @@
-enum WifiPermissionStatus {
-  enabled,
-  notConfigured,
-  notSupported,
-}
-
-class WifiPermissionResult {
-  final WifiPermissionStatus status;
-  final String message;
-
-  const WifiPermissionResult({
-    required this.status,
-    required this.message,
-  });
-}
-
-class WifiAwarePermissionResult {
-  final WifiPermissionStatus status;
-  final String message;
-
-  const WifiAwarePermissionResult({
-    required this.status,
-    required this.message,
-  });
-}
+import 'wifi_permissions_types.dart';
 
 /// Check if WiFi permissions are properly configured for Ditto
 Future<WifiPermissionResult> checkWifiPermissions() async {
